// frontend-angular/src/app/models/user.ts
export interface User {
    id: number;
  firstName: string;
  lastName: string;
  position: string;
  companyName: string;
  officePhone: string;
  mobilePhone: string;
  address: string;
  websiteUrl: string;
  businessEmail: string;
  logoUrl: string;
  photoUrl: string;
  linkedInUrl: string;
  youtubeUrl: string;
  instagramUrl: string;
  email: string;
  password?: string;
  role: number;
  departmentId: number;
}